split schema update to its own class
