+++ settings/command_list.json
@@ -14,7 +14,7 @@
             "description": "Prints all the commands"
         }
     },
-    "apiUrlInterpreter": {
+    "apiSchema": {
         "updateSchema": {
             "aliases": [
                 "update_schema",
