+++ src/ApiUrlInterpreter.ts
@@ -1,97 +1,59 @@
 import Discord = require("discord.js");
-import prettyMs = require("pretty-ms");
-import { Response } from "node-fetch";
-import fetch from "node-fetch";
 import fs = require("fs");
+import fetch from "node-fetch";
+import prettyMs = require("pretty-ms");
 
+import { APISchema, Path } from "./ApiSchema";
 import { fileBackedObject } from "./FileBackedObject";
-import { SharedSettings } from "./SharedSettings";
-import { PersonalSettings } from "./PersonalSettings";
-import { setTimeout, clearTimeout } from "timers";
-import { platform } from "os";
 import levenshteinDistance from "./LevenshteinDistance";
+import { PersonalSettings, SharedSettings } from "./SharedSettings";
+
 import { ENODATA } from "constants";
-
-class PathRegexCollection {
-    constructor(validString: string, invalidString: string) {
-        this.valid = validString;
-        this.invalid = invalidString;
-    }
-
-    public invalid: string;
-    public valid: string;
-};
-
-class SchemaRegexCollection extends PathRegexCollection {
-    constructor(name: string, schema: any, validString: string, invalidString: string) {
-        super(validString, invalidString);
-
-        this.name = name;
-        this.schema = schema;
-    }
-    
-    public name: string;
-    public schema: any;
-};
+import { Response } from "node-fetch";
 
 class RatelimitResult {
-    constructor(rateLimit: number, startTime: number|null) {
+    public rateLimit: number;
+    public startTime: number | null;
+
+    constructor(rateLimit: number, startTime: number | null) {
         this.rateLimit = rateLimit;
         this.startTime = startTime;
     }
-
-    public rateLimit: number;
-    public startTime: number|null;
-};
-
-class Path {
-    public methodType: "GET"|"POST";
-    public regex: PathRegexCollection;
-    public parameterInfo: SchemaRegexCollection[];
-    public method: string;
-    public name: string;
-};
+}
 
 export default class ApiUrlInterpreter {
     private static ratelimitErrorMs = 100;
-    
+
     private bot: Discord.Client;
     private sharedSettings: SharedSettings;
     private personalSettings: PersonalSettings;
-    private timeOut: NodeJS.Timer|null;
+    private apiSchema: APISchema;
     private iterator: number = 1;
 
-    private baseUrl: string;
-    private platforms: string[];
-    private platformRegexString: string;
-
     private applicationRatelimitLastTime: number = 0;
-    private methodRatelimitLastTime: {[method: string]: number} = {};
+    private methodRatelimitLastTime: { [method: string]: number } = {};
     private applicationStartTime: number = 0;
-    private methodStartTime: {[method: string]: number} = {};
-
-    private paths: Path[] = [];
+    private methodStartTime: { [method: string]: number } = {};
 
     private fetchSettings: Object;
 
-    constructor(bot: Discord.Client, sharedSettings: SharedSettings) {
+    constructor(bot: Discord.Client, sharedSettings: SharedSettings, apiSchema: APISchema) {
         console.log("Requested API URL Interpreter extension..");
 
         this.sharedSettings = sharedSettings;
         this.personalSettings = sharedSettings.botty;
+        this.apiSchema = apiSchema;
         console.log("Successfully loaded API URL Interpreter settings.");
 
-        this.fetchSettings = { 
-            headers: { 
-                "X-Riot-Token": this.personalSettings.riotApi.key 
+        this.fetchSettings = {
+            headers: {
+                "X-Riot-Token": this.personalSettings.riotApi.key
             }
         };
 
         this.bot = bot;
         this.bot.on("ready", this.onBot.bind(this));
         this.bot.on("message", this.onMessage.bind(this));
-
-        this.updateSchema();
     }
 
     onBot() {
@@ -99,20 +61,19 @@ export default class ApiUrlInterpreter {
     }
 
     onMessage(message: Discord.Message) {
-        if (message.author.bot) return; 
+        if (message.author.bot) return;
 
         this.onRiotApiURL(message);
     }
 
-    async onRiotApiURL(message: Discord.Message, content: string|null = null) {
-        
+    async onRiotApiURL(message: Discord.Message, content: string | null = null) {
+
         // Init message if missing, also append with space.
         if (!content) content = message.content.replace(/(`){1,3}(.*?)(`){1,3}/g, "") + " ";
 
         if (content.indexOf("https://") == -1) return; // We're about to do ~80 regex tests, better make sure that it's on a message with a URL
 
-        for (let i = 0; i < this.paths.length; i++) {
-            const path = this.paths[i];
+        for (const path of this.apiSchema.paths) {
 
             // Check if path was valid
             const validMatch = new RegExp(path.regex.valid, "g").exec(content);
@@ -124,16 +85,14 @@ export default class ApiUrlInterpreter {
                 const replyMessage = Array.isArray(replyMessages) ? replyMessages[0] : replyMessages;
 
                 const region = validMatch[1];
-                
+
                 // Path works fine, make a request
                 await this.makeRequest(path, region, validMatch[0], replyMessage);
                 return;
             }
         }
-        
-        for (let i = 0; i < this.paths.length; i++) {
-            const path = this.paths[i];
-            
+
+        for (const path of this.apiSchema.paths) {
             // Now check if it's the same path, but with incorrect parameters.
             const invalidMatch = new RegExp(path.regex.invalid, "g").exec(content);
             if (invalidMatch && invalidMatch.length > 0) {
@@ -145,14 +104,14 @@ export default class ApiUrlInterpreter {
                 const closestPlatform = this.getClosestPlatform(invalidMatch[1]);
                 if (closestPlatform) {
                     errorIdentified = true;
-                    mistakes.push(`- The platform \`${invalidMatch[1]}\` is invalid, did you mean: \`${closestPlatform}\`? Expected one of the following values: \`${this.platforms.join(", ")}\``);
+                    mistakes.push(`- The platform \`${invalidMatch[1]}\` is invalid, did you mean: \`${closestPlatform}\`? Expected one of the following values: \`${this.apiSchema.platforms.join(", ")}\``);
                 }
 
                 invalidMatch.splice(0, 2); // Remove url and platform from array
-                
+
                 // This now contains all required parameters.
                 for (let j = 0; j < Math.max(path.parameterInfo.length, invalidMatch.length); j++) {
-                    
+
                     const parameter = path.parameterInfo[j]; // All info about this parameter
                     const value = invalidMatch[j];
 
@@ -170,7 +129,7 @@ export default class ApiUrlInterpreter {
                     if (parameter.schema.enum) { // If there are limited options what it could be, show them
                         type = parameter.schema.enum.join("/");
                     }
-                    
+
                     mistakes.push(`- Parameter ${j + 1}, expected \`${parameter.name} (${type})\` right after \`${location}\`, got "${value}".`);
                     errorIdentified = true;
                 }
@@ -189,39 +148,20 @@ export default class ApiUrlInterpreter {
     }
 
     getClosestPlatform(platform: string) {
-        const validPlatform = new RegExp(this.platformRegexString, "g").exec(platform);
+        const validPlatform = new RegExp(this.apiSchema.platformRegexString, "g").exec(platform);
         if (validPlatform) return null;
-        
-        return this.platforms.map(p => { 
-            return { 
-                platform: p, 
-                distance: levenshteinDistance(platform, p) 
+
+        return this.apiSchema.platforms.map(p => {
+            return {
+                platform: p,
+                distance: levenshteinDistance(platform, p)
             }
         })
-        .sort((a, b) => a.distance - b.distance)[0].platform;
-    }
-
-    public async onUpdateSchemaRequest(message: Discord.Message, isAdmin: boolean, command: string, args: string[]) {
-        const replyMessagePromise = message.channel.send("Updating schema..");
-
-        console.log(`${message.author.username} requested a schema update.`);
-        await this.updateSchema();
-
-        const newMessage = "Updated schema.";
-        let replyMessage = await replyMessagePromise;
-
-        // Could be an array? Would be weird.
-        if (Array.isArray(replyMessage)) {
-            console.warn("replyMessage is an array, what do you know?");
-            for (let i = 0; i < replyMessage.length; i++) {
-                replyMessage[i].edit(newMessage);
-            }
-        }
-        else replyMessage.edit(newMessage);
+            .sort((a, b) => a.distance - b.distance)[0].platform;
     }
 
     async makeRequest(path: Path, region: string, url: string, message: Discord.Message) {
-        
+
         const currentTime = Date.now();
         if (currentTime < this.applicationRatelimitLastTime) {
             const timeDiff = prettyMs(this.applicationRatelimitLastTime - currentTime, { verbose: true });
@@ -235,7 +175,7 @@ export default class ApiUrlInterpreter {
             message.edit(`We are ratelimited by the method (${servicedMethodName}), please wait ${timeDiff}.`);
             return;
         }
-        
+
         try {
             const resp = await fetch(url, this.fetchSettings);
             this.handleResponse(resp, message, url, servicedMethodName);
@@ -249,7 +189,7 @@ export default class ApiUrlInterpreter {
         if (resp === null) {
             console.warn(`Not handling ratelimits due to missing response.`);
             return;
-        } 
+        }
 
         // Set start times
         if (this.applicationStartTime === 0) {
@@ -265,11 +205,11 @@ export default class ApiUrlInterpreter {
             const limitHeader = resp.headers.get("x-app-rate-limit");
 
             if (countHeader && limitHeader) {
-                
+
                 const appCountStrings = countHeader.split(",");
                 const appLimitStrings = limitHeader.split(",");
                 const appResult = this.handleRatelimit("application", servicedMethodName, this.applicationStartTime, appCountStrings, appLimitStrings);
-    
+
                 if (appResult) {
                     this.applicationRatelimitLastTime = appResult.rateLimit;
                     if (appResult.startTime !== null) this.applicationStartTime = appResult.startTime;
@@ -311,8 +251,7 @@ export default class ApiUrlInterpreter {
                 result: await resp.json(),
             };
 
-            fs.writeFile(localFile, JSON.stringify(json), null, (err: NodeJS.ErrnoException) =>
-            {
+            fs.writeFile(localFile, JSON.stringify(json), null, (err: NodeJS.ErrnoException) => {
                 if (err != null) {
                     message.edit(`Woah, something went wrong trying to fetch ${url}, sorry!`);
                     console.warn(`Error ${err.code} (${err.name}) while trying to save \`${url}\` to \`${localFile}\`: ${err.message}`);
@@ -332,19 +271,19 @@ export default class ApiUrlInterpreter {
         }
     }
 
-    handleRatelimit(ratelimitType: string, methodName: string, startTime: number, countStrings: string[], limitStrings: string[]): RatelimitResult|null {
-        
+    handleRatelimit(ratelimitType: string, methodName: string, startTime: number, countStrings: string[], limitStrings: string[]): RatelimitResult | null {
+
         let found = false;
         let longestSpreadTime = 0;
-        let resultStartTime: number|null = 0;
+        let resultStartTime: number | null = 0;
         let resultRatelimit = 0;
 
         for (let i = 0; i < countStrings.length; i++) {
             const splitCount = countStrings[i].split(":");
             const count = parseInt(splitCount[0].trim());
             const time = parseInt(splitCount[1].trim());
-            
-            const limit = limitStrings.find(function(element) {
+
+            const limit = limitStrings.find(function (element) {
                 return element.indexOf(`:${time}`) != -1;
             });
             if (!limit) {
@@ -375,133 +314,5 @@ export default class ApiUrlInterpreter {
 
         if (found === false) return null;
         return new RatelimitResult(resultRatelimit, resultStartTime)
-    }
-
-    async updateSchema() {
-        try { 
-            const response = await fetch(`http://www.mingweisamuel.com/riotapi-schema/openapi-3.0.0.json`, {
-                method: "GET",
-                headers: {
-                    Accept: "application/json",
-                    "Content-Type": "application/json"
-                }
-            });
-
-            if (response.status !== 200) {
-                console.error("HTTP Error trying to get schema: " + response.status);
-                return;
-            }
-
-            let schema = await response.json();
-
-            let baseUrlRegex = this.constructBaseUrlRegex(schema);
-            let invalidBaseUrlRegex = this.constructInvalidBaseUrlRegex(schema);
-
-            this.paths = [];
-            
-            for (const pathName in schema.paths) {
-                const pathSchema = schema.paths[pathName];
-                const methodSchema = pathSchema.get ? pathSchema.get : pathSchema.post;
-
-                if (!methodSchema) continue; // Only handle GET/POST
-                if (methodSchema.operationId.startsWith("tournament")) continue;
-
-                let path = new Path();
-                path.methodType = pathSchema.get ? "GET" : "POST";
-                path.method = methodSchema.operationId;
-                this.constructRegex(path, baseUrlRegex, invalidBaseUrlRegex, pathName, methodSchema);
-                console.assert(path.regex, "Path Regex should be set");
-                
-                this.paths.push(path);
-            }
-
-            // This fixes the issue where it would match getAllChampionsMasteries before a specific champion mastery (which starts the same but has extra parameters)
-            this.paths = this.paths.sort((a, b) => b.name.length - a.name.length);
-        }
-        catch (e) {
-            console.error("Schema fetch error: " + e.message);
-        }
-
-        if (this.timeOut) {
-            clearTimeout(this.timeOut);
-            this.timeOut = null;
-        }
-        this.timeOut = setTimeout(this.updateSchema.bind(this), this.sharedSettings.apiUrlInterpreter.timeOutDuration);
-    }
-
-    constructBaseUrlRegex(schema: any): string {
-        let baseUrl = this.escapeRegex(schema.servers[0].url);
-        this.platforms = schema.servers[0].variables.platform.enum;
-
-        // This makes a regex string from all the platform options
-        this.platformRegexString = `(${this.platforms.join("|")})`;
-
-        return baseUrl.replace(/{platform}/g, this.platformRegexString);
-    }
-
-    constructInvalidBaseUrlRegex(schema: any): string {
-        let baseUrl = this.escapeRegex(schema.servers[0].url);
-        return baseUrl.replace(/{platform}/g, "(.*?)");
-    }
-
-    constructRegex(path: Path, validBase: string, invalidBase: string, pathName: string, methodSchema: any) {
-        
-        path.name = pathName;
-        path.parameterInfo = [ ];
-
-        let invalidPath = invalidBase + this.escapeRegex(pathName);
-        let validPath = validBase + this.escapeRegex(pathName);
-
-        if (!methodSchema.parameters) {
-            path.regex = new PathRegexCollection(validPath, invalidPath);
-            return;
-        }
-
-        const invalidWith = "([^\\s^\\/]*)";
-        for (let i = 0; i < methodSchema.parameters.length; i++) {
-
-            const parameter = methodSchema.parameters[i];
-            if (parameter.required == false) continue;
-
-            const parameterReplace = new RegExp(`{${parameter.name}}`, "g");
-
-            let validWith = "([^\\s^\\/]+)";
-
-            if (parameter.schema.enum) {
-                validWith = `(${parameter.schema.enum.join("|")})`;
-            }
-            else switch(parameter.schema.type) {
-                default:
-                    console.warn(`Unhandled schema parameter in ${pathName}: ${parameter.name} is a ${parameter.schema.type}`);
-                    break;
-
-                case "string":
-                    break;
-
-                case "integer":
-                    validWith = "([0-9]+)";
-                    break;
-            }
-
-            invalidPath = invalidPath.replace(parameterReplace, invalidWith);
-            validPath = validPath.replace(parameterReplace, validWith);
-
-            path.parameterInfo.push(new SchemaRegexCollection(parameter.name, parameter.schema, validWith, invalidWith));
-        }
-
-        validPath += "\\/?"; // Allow urls to end with a trailing /
-        validPath += "\\s"; // Message will always end in a whitespace, use this a delimiter at the end of valid paths
-
-        // If the last parameter is missing from the url, don't require the last / match for invalids.
-        const lastIndex = invalidPath.lastIndexOf("\\/" + invalidWith);
-        if (lastIndex !== -1) {
-            invalidPath = invalidPath.substr(0, lastIndex) + "\\/?" + invalidWith;
-        }
-
-        path.regex = new PathRegexCollection(validPath, invalidPath);
-    }
-    
-    escapeRegex(regex: string): string {
-        return regex.replace(/[\-\[\]\/\(\)\*\+\?\.\\\^\$\|]/g, "\\$&");
     }
 }
